refactor card picker
move cardPickerState to separate file
